+++ packages/test-ui/src/components/Router.tsx
@@ -34,7 +34,7 @@ const LOCAL_CHAINS = [1337, 1338];
 
 const TESTNET_CHAINS = [421611, 97, 43113, 5, 42, 80001, 4, 3];
 
-const MAINNET_CHAINS = [56, 100, 137, 250, 1285, 42161, 43114];
+const MAINNET_CHAINS = [10, 56, 100, 137, 250, 1285, 42161, 43114];
 
 type BalanceEntry = {
   chain: string;