+++ CHANGELOG.md
@@ -2,7 +2,15 @@
 
 ## Next Release
 
+- [sdk] Gelato support for Arbitrum
+- [subgraph] subgraph lite for routers and sdk
+- [subgraph] add sending tx count for explorer
+- [router] update active transactions logs!minor
+
+## 0.0.80
+
 - [utils] Enforce minimum subgraph sync buffer
+- [utils] Subgraph starts as synced
 
 ## 0.0.79